update monorepo script

+++ scripts/monorepo.sh
@@ -24,15 +24,17 @@ github_org=LedgerHQ
 # Repositories considered as "apps"
 apps="ledger-live-desktop ledger-live-mobile"
 # Repositories considered as "libs"
-libs="ledger-live-common ledgerjs"
+libs="ledger-live-common ledgerjs ui"
 # Main git branch
-main_branch=master
+main_branch=main
 
 ## Constants
 root_dir=`pwd` # The root dir
 apps_dir=apps  # Will put the apps inside this folder
 libs_dir=libs  # Will put the libs inside this folder
 skip_bootstrap=$NO_BOOTSTRAP # Skip bootstrap, use this if the target folder already exists
+skip_tags=$NO_TAGS # Prevents fetching and merging tags
+only_main_branch=$ONLY_MAIN_BRANCH
 
 ## Checks
 
@@ -71,24 +73,53 @@ else
   echo -e "\033[1;33m[Boostrap] <SKIPPED> \033[0;0m"
 fi
 
-# This directory will contain all final tag refs (namespaced)
-rm -Rf .git/refs/namespaced-tags
-if [ -e ".git/refs/tags" ]; then
-  mv .git/refs/tags .git/refs/namespaced-tags
-else
-  mkdir -p .git/refs/namespaced-tags
+if [ -z "$skip_tags" ]; then
+  # This directory will contain all final tag refs (namespaced)
+  rm -Rf .git/refs/namespaced-tags
+  if [ -e ".git/refs/tags" ]; then
+    mv .git/refs/tags .git/refs/namespaced-tags
+  else
+    mkdir -p .git/refs/namespaced-tags
+  fi
 fi
 
 ## Repo. migration
 
+function merge_branch {
+  local remote=$1
+  local remote_branch=$2
+  local local_branch=$3
+  local target_dir=$4
+
+  echo -e "  > Merging branch $repo/$remote_branch into $local_branch"
+  if git rev-parse -q --verify "$local_branch"; then
+    # Branch already exists, just check it out (and clean up the working dir)
+    git checkout -q -b "$local_branch" "$remote_branch"
+    git checkout -q -- .
+    git clean -f -d
+  else
+    # Create a fresh branch with an empty root commit"
+    git checkout -q --orphan "$local_branch"
+    # The ignore unmatch is necessary when this was a fresh repo
+    git rm -rfq --ignore-unmatch .
+    git commit -q --allow-empty -m "monorepo: init $repo:$remote_branch"
+  fi
+  # See: https://docs.github.com/en/get-started/using-git/about-git-subtree-merges
+  (
+    git merge -q --no-commit -s ours "$repo/$remote_branch" --allow-unrelated-histories &&\
+    git read-tree --prefix="$target_dir/$repo/" -u "$repo/$remote_branch" &&\
+    git commit -q --no-verify --allow-empty -m "monorepo: merge $repo:$remote_branch"
+  ) || true
+}
+
 function migrate_repo {
-  repo=$1
+  local repo=$1
   if [ $2 = 1 ]; then
     echo -e "\033[1m[Migration](App: $repo) \033[0;0m"
-    target_dir=$apps_dir
+    local target_dir=$apps_dir
   else
     echo -e "\033[1m[Migration](Lib: $repo) \033[0;0m"
-    target_dir=$libs_dir
+    local target_dir=$libs_dir
   fi
 
   remote_exists=`git remote | grep "^$repo$" || echo ''`
@@ -99,35 +130,28 @@ function migrate_repo {
     echo -e "  > Remote exists"
   fi
   echo -e "  > Fetching remote"
-  git fetch -q "$repo"
 
-  # Now we've got all tags in .git/refs/tags: put them away for a sec
-  echo -e "  > Namespace and backup tags"
-  if [[ -n "$(ls .git/refs/tags)" ]]; then
-    mv .git/refs/tags ".git/refs/namespaced-tags/$repo"
+  if [ -z "$skip_tags" ]; then
+    git fetch -q "$repo"
+    # Now we've got all tags in .git/refs/tags: put them away for a sec
+    echo -e "  > Namespace and backup tags"
+    if [[ -n "$(ls .git/refs/tags)" ]]; then
+      mv .git/refs/tags ".git/refs/namespaced-tags/$repo"
+    fi
+  else
+    git fetch -q -n "$repo"
   fi
+
 
   branches=`git branch -r --list --format="%(refname:lstrip=3)" "$repo/*"`
   for branch in $branches; do
-    echo -e "  > Merging branch $branch"
-    if git rev-parse -q --verify "$branch"; then
-      # Branch already exists, just check it out (and clean up the working dir)
-      git checkout -q "$branch"
-      git checkout -q -- .
-      git clean -f -d
-    else
-      # Create a fresh branch with an empty root commit"
-      git checkout -q --orphan "$branch"
-      # The ignore unmatch is necessary when this was a fresh repo
-      git rm -rfq --ignore-unmatch .
-      git commit -q --allow-empty -m "monorepo: init $repo:$branch"
+    if [ -z "$only_main_branch" ]; then
+      # Merge every branch.
+      merge_branch $repo $branch $branch $target_dir
+    elif [[ $branch = $main_branch || $branch = "master" ]]; then
+      # Merge only master & main branches.
+      merge_branch $repo $branch $main_branch $target_dir
     fi
-    # See: https://docs.github.com/en/get-started/using-git/about-git-subtree-merges
-    (
-      git merge -q --no-commit -s ours "$repo/$branch" --allow-unrelated-histories &&\
-      git read-tree --prefix="$target_dir/$repo/" -u "$repo/$branch" &&\
-      git commit -q --no-verify --allow-empty -m "monorepo: merge $repo:$branch"
-    ) || true
   done
 
   echo -e "  > Done"
@@ -152,9 +176,11 @@ echo -e "\033[1m\
 \033[0;0m"
 
 # Restore all namespaced and backuped tags
-echo -e "  > Restore tags"
-rm -rf .git/refs/tags
-mv .git/refs/namespaced-tags .git/refs/tags
+if [ -z "$skip_tags" ]; then
+  echo -e "  > Restore tags"
+  rm -rf .git/refs/tags
+  mv .git/refs/namespaced-tags .git/refs/tags
+fi
 
 echo -e "  > Checkout $main_branch"
 git checkout -q $main_branch
